requirements.txt
asgiref==3.2.3
asn1crypto==0.24.0
bcrypt==3.1.6
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
configparser==5.0.0
cryptography==2.6.1
da-vinci==0.2.2
distro==1.3.0
Django==3.0exit.4
django-adminrestrict==2.0.2
django-bootstrap-form==3.4
django-jet==1.0.7
django-jet2==1.0.12
django-recaptcha==2.0.6
dnspython==1.16.0
docopt==0.6.2
easy-thumbnails==2.7
entrypoints==0.3
fasteners==0.12.0
filelock==3.0.12
future==0.16.0
google-api-python-client==1.4.1
httplib2==0.11.3
idna==2.8
keyring==18.0.1
keyrings.alt==3.1.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.2.3
Mako==1.0.7
MarkupSafe==1.1.0
monotonic==1.5
mysql-client==0.0.1
oauth2client==4.1.3
oauthlib==2.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==6.1.0
pinax-calendars==3.0.0
pinax-templates==2.0.6
protobuf==3.6.1
protonvpn-cli==2.2.2
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-dateutil==2.7.3
python-debian==0.1.36
python-decouple==3.3
pythondialog==3.5.1
pytz==2019.2
pyxdg==0.25
PyYAML==5.1.2
qrcode==6.1
redis==3.5.3
reportlab==3.5.23
requests==2.21.0
requests-unixsocket==0.1.5
rsa==4.6
SecretStorage==2.3.1
shortuuid==1.0.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.1
twilio==6.44.1
uritemplate==3.0.1
urllib3==1.24.1
validate-email==1.3
wadllib==1.3.3
